# Chapter 8/server/API.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: "RESTful API Suwit Games"
  description: "This RESTful API is being used to manage the suwit games"
servers:
  - url: /
    description: "Mock Server for Suwit Games"
...
paths:
...
  /v1/login:
    post:
      summary: Retieve access token user
      description: This endpoint to retrieve the access token when user logged in with correct credential.
      operationId: userLogin
      tags:
        - User & Auth
      requestBody:
        $ref: "#/components/requestBodies/userLogin"
      responses:
        200:
          $ref: "#/components/responses/isLoginSuccess"
        400:
          $ref: "#/components/responses/isLoginFailed"
...
  /v1/profile:
    get:
      summary: Retrieve user profile
      description: This endpoint to get the user profile who logged in
      operationId: getUser
      tags:
        - User
        - Biodata
      responses:
        200:
          $ref: "#/components/responses/showProfile"
        404:
          $ref: "#/components/responses/errorResponse"
    post:
      summary: Create biodata
      description: This endpoint to create biodata for user who logged in
      operationId: createBiodata
      tags:
        - User
        - Biodata
      requestBody:
        $ref: "#/components/requestBodies/createBiodata"
      responses:
        201:
          $ref: "#/components/responses/isCreateBiodataSuccess"
        400:
          $ref: "#/components/responses/isCreateBiodataFailed"
    put:
      summary: Update biodata
      description: This endpoint to update biodata for user who logged in
      operationId: updateBiodata
      tags:
        - User
        - Biodata
      requestBody:
        $ref: "#/components/requestBodies/updateBiodata"
      responses:
        201:
          $ref: "#/components/responses/isUpdateBiodataSuccess"
        400:
          $ref: "#/components/responses/isUpdateBiodataFailed"
...
  /v1/games:
    get:
      summary: Retrieve information game
      description: This endpoint to retrieve all games information
      operationId: getGames
      tags:
        - Game
      responses:
        200:
          $ref: "#/components/responses/getGames"
        400:
          $ref: "#/components/responses/errorResponse"
    post:
      summary: Create Games
      description: This endpoint to create games
      operationId: createGames
      tags:
        - Game
      responses:
        201:
          $ref: "#/components/responses/successCreateGames"
        400:
          $ref: "#/components/responses/errorResponse"
...
components:
  requestBodies:
...
    userLogin:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - username
              - password
            properties:
              username:
                type: string
                example: daffa404
              password:
                type: string
                example: thisIsNotPassword
...
    createBiodata:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - "name"
              - "phone"
              - "address"
              - "avatar"
              - "bio"
            $ref: "#/components/schemas/biodataEntity"
    updateBiodata:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - "name"
              - "phone"
              - "address"
              - "avatar"
              - "bio"
            $ref: "#/components/schemas/updateBiodataEntity"
...
  responses:
...
    errorResponse:
      description: Response when something error
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: FAIL
              data:
                type: object
                properties:
                  message:
                    type: string
                    example: Something error
...
    isLoginSuccess:
      description: response if user login with correct credential.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK
              data:
                $ref: "#/components/schemas/getTokenSchema"
    isLoginFailed:
      description: response when the user login with invalid credential.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: FAIL
              data:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid username or password!
...
    showProfile:
      description: response details of user who logged in
      content:
        application/json:
          schema:
            type: object
            properties:
              user:
                type: object
                properties:
                  id:
                    type: string
                    example: 8bbefc02-6889-46c0-92da-4dadde73e4ee
                  username:
                    type: string
                    example: daffa404
                  email:
                    type: string
                    example: daffa@suwitgames.com
                  created_at:
                    type: string
                    example: 28-10-2021 23:30 +07:00
                  updated_at:
                    type: string
                    example: 28-10-2021 23:30 +07:00
                  biodata:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 4eb527a7-db0e-42f9-9e57-21422a01619b
                      name:
                        type: string
                        example: Daffa Hanifisyafiq
                      phone:
                        type: string
                        example: +62 877-1234-9999
                      address:
                        type: string
                        example: Jalan Kenangan Gg. Rindu No. 28
                      avatar:
                        type: string
                        example: suwitgames.com/daffa404.png
                      bio:
                        type: string
                        example: miss someone far away
                      created_at:
                        type: string
                        example: 28-10-2021 23:30 +07:00
                      updated_at:
                        type: string
                        example: 28-10-2021 23:30 +07:00
                      userId:
                        type: string
                        example: 8bbefc02-6889-46c0-92da-4dadde73e4ee
    isCreateBiodataSuccess:
      description: response create biodata of user who logged in
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK
              data:
                type: object
                properties:
                  id:
                    type: string
                    example: 4eb527a7-db0e-42f9-9e57-21422a01619b
                  name:
                    type: string
                    example: Daffa Hanifisyafiq
                  phone:
                    type: string
                    example: +62 877-1234-9999
                  address:
                    type: string
                    example: Jalan Kenangan Gg. Rindu No. 28
                  avatar:
                    type: string
                    example: suwitgames.com/daffa404.png
                  bio:
                    type: string
                    example: miss someone far away
                  created_at:
                    type: string
                    example: 28-10-2021 23:30 +07:00
                  updated_at:
                    type: string
                    example: 28-10-2021 23:30 +07:00
    isCreateBiodataFailed:
      description: response when invalid request.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: FAIL
              data:
                type: object
                properties:
                  message:
                    type: string
                    example: Create biodata failed, please fill field correctly
    isUpdateBiodataSuccess:
      description: response update biodata of user who logged in
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK
              data:
                type: object
                properties:
                  id:
                    type: string
                    example: 4eb527a7-db0e-42f9-9e57-21422a01619b
                  name:
                    type: string
                    example: Daffa
                  phone:
                    type: string
                    example: +62 877-1234-1234
                  address:
                    type: string
                    example: Jalan Kenangan Gg. Rindu No. 28
                  avatar:
                    type: string
                    example: suwitgames.com/daffa404.png
                  bio:
                    type: string
                    example: miss someone far away
                  created_at:
                    type: string
                    example: 28-10-2021 23:30 +07:00
                  updated_at:
                    type: string
                    example: 28-10-2021 23:30 +07:00
    isUpdateBiodataFailed:
      description: response when invalid request.
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: FAIL
              data:
                type: object
                properties:
                  message:
                    type: string
                    example: Update biodata failed, please fill field correctly
...
    getGames:
      description: Response for /games endpoint
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK
              data:
                type: object
                properties:
                  id:
                    type: string
                    example: 5380cc52-2508-4a7c-aba6-59046518c58b
                  played_at:
                    type: string
                    example: 01-12-2021 00:01 +07:00
                  ended_at:
                    type: string
                    example: 04-12-2021 00:01 +07:00
                  isActive:
                    type: boolean
                    example: true
                  isWinner:
                    type: string
                    example: daffa404
                  roundActive:
                    type: string
                    example: "3"
                  created_at:
                    type: string
                    example: 01-12-2021 00:01 +07:00
                  updated_at:
                    type: string
                    example: 04-12-2021 00:01 +07:00
    successCreateGames:
      description: Response for create gamescontent
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: OK
              data:
                type: object
                properties:
                  played_at:
                    type: string
                    example: 01-12-2021 00:01 +07:00
                  ended_at:
                    type: string
                    example: 04-12-2021 00:01 +07:00
                  isActive:
                    type: boolean
                    example: true
                  isWinner:
                    type: string
                    example: daffa404
                  roundActive:
                    type: string
                    example: "3"
                  created_at:
                    type: string
                    example: 01-12-2021 00:01 +07:00
                  updated_at:
                    type: string
                    example: 04-12-2021 00:01 +07:00
...
  schemas:
    getTokenSchema:
      type: object
      properties:
        message:
          type: string
          example: Login success!
        accessToken:
          type: string
          example: XXXXX
        refreshToken:
          type: string
          example: XXXXX
    biodataEntity:
      type: object
      properties:
        name:
          type: string
          example: Daffa Hanifisyafiq
        phone:
          type: string
          example: +62 877-1234-9999
        address:
          type: string
          example: Jalan Kenangan Gg. Rindu No. 28
        avatar:
          type: string
          example: suwitgames.com/daffa404.png
        bio:
          type: string
          example: miss someone far away
    updateBiodataEntity:
      type: object
      properties:
        name:
          type: string
          example: Daffa
        phone:
          type: string
          example: +62 877-1234-1234
        address:
          type: string
          example: Jalan Kenangan Gg. Rindu No. 28
        avatar:
          type: string
          example: suwitgames.com/daffa404.png
        bio:
          type: string
          example: miss someone far away
...